+++ src/Core/Schema/Schema.php
@@ -37,6 +37,7 @@ use LaravelJsonApi\Core\Support\Arr;
 use LaravelJsonApi\Core\Support\Str;
 use LogicException;
 use function array_keys;
+use function iterator_to_array;
 use function sprintf;
 
 abstract class Schema implements SchemaContract, IteratorAggregate
@@ -464,23 +465,35 @@ abstract class Schema implements SchemaContract, IteratorAggregate
      */
     public function sortFields(): iterable
     {
-        $id = $this->id();
+        return array_keys(iterator_to_array(
+            $this->allSortFields()
+        ));
+    }
 
-        if ($id->isSortable()) {
-            yield $id->name();
-        }
-
-        /** @var Attribute $attr */
-        foreach ($this->attributes() as $attr) {
-            if ($attr->isSortable()) {
-                yield $attr->name();
+    /**
+     * @inheritDoc
+     */
+    public function sortField(string $name)
+    {
+        foreach ($this->allSortFields() as $key => $sortable) {
+            if ($name === $key) {
+                return $sortable;
             }
         }
 
-        /** @var Sortable $sortable */
-        foreach ($this->sortables() as $sortable) {
-            yield $sortable->sortField();
-        }
+        throw new LogicException(sprintf(
+            'Sort field %s does not exist on schema %s.',
+            $name,
+            get_class($this),
+        ));
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function sortables(): iterable
+    {
+        return [];
     }
 
     /**
@@ -489,16 +502,6 @@ abstract class Schema implements SchemaContract, IteratorAggregate
     public function authorizable(): bool
     {
         return true;
-    }
-
-    /**
-     * Get additional sortable fields.
-     *
-     * @return iterable
-     */
-    protected function sortables(): iterable
-    {
-        return [];
     }
 
     /**
@@ -545,5 +548,31 @@ abstract class Schema implements SchemaContract, IteratorAggregate
         return $this->relations = collect($this->allFields())
             ->whereInstanceOf(Relation::class)
             ->all();
+    }
+
+    /**
+     * Iterate through all the sort fields.
+     *
+     * @return iterable
+     */
+    private function allSortFields(): iterable
+    {
+        $id = $this->id();
+
+        if ($id->isSortable()) {
+            yield $id->name() => $id;
+        }
+
+        /** @var Attribute $attribute */
+        foreach ($this->attributes() as $attribute) {
+            if ($attribute->isSortable()) {
+                yield $attribute->name() => $attribute;
+            }
+        }
+
+        /** @var Sortable $sortable */
+        foreach ($this->sortables() as $sortable) {
+            yield $sortable->sortField() => $sortable;
+        }
     }
 }